feat(code): Add message builder context
Add context to message builders.

This change breaks the 2nd `format()` argument, it now expects an Object instead of a function. This should be released with a new major, but given the original version was only released a few days ago, and this change is unlikely to impact any of the tiny amount of downloads in the past few days, this will be released using only a minor semver increment. After this change, stricter semver releases should be adhered too.