feat(code): Factory Methods
Change casing of factory methods.

This is a breaking change, but as this package was only published a few days ago and the fact that Factory methods are mostly internal (and unlikely used by other packages) release as a minor change. Stricter semver will be followed after the release of react-intl-fmt.