feat(code): Rename IntlMessageFormat
Rename all usages of `IntlMessageFormat` to `TagMessageFormat`.

This should only impact browser usage - and because it’s not currently being published to a CDN (and it was only first released a few days ago) this will be released only as a minor change.

This package will follow semver after `react-intl-fmt` is released